One error arise when the precipitation probability is equal to 1 (100%). jsonDecode interprets it as integer, raising a eror. Properly exception manage.

+++ lib/features/booking/infrastructure/dto/daily_weather_mode.dart
@@ -10,7 +10,7 @@ class Daily {
 
   Daily.fromJson(Map<String, dynamic> json) {
     dt = json['dt'];
-    pop = json['pop'];
+    pop = json['pop'].toDouble();
   }
 
   Map<String, dynamic> toJson() {
